+++ src/components/header.tsx
@@ -9,6 +9,7 @@ import * as colors from '../common/colors';
 const HeaderWrapper = styled.div`
     position: sticky;
     top: 0;
+    z-index: 2;
     background: ${colors.charcoal};
     color: ${colors.white};
     margin-bottom: '1.45rem';
